+++ src/Controller/VesselsController.php
@@ -19,6 +19,8 @@ class VesselsController extends AppController {
 	public function initialize() {
         parent::initialize();
         $this->loadComponent('RequestHandler');
+		$this->response->header('Access-Control-Allow-Origin', '*');
+		$this->response->header('Access-Control-Allow-Methods', 'POST, GET, OPTIONS, PUT, DELETE');        
     }
 
 /**
@@ -29,7 +31,7 @@ class VesselsController extends AppController {
 
 	public function index() {
 		$this->paginate = [
-			'contain' => ['Creators', 'Modifiers', 'VesselOwners']
+			'contain' => ['VesselOwners']
 		];
 		$this->set('vessels', $this->paginate($this->Vessels));
         $this->set('_serialize', ['vessels']);
@@ -43,10 +45,19 @@ class VesselsController extends AppController {
  * @throws \Cake\Network\Exception\NotFoundException
  */
 	public function view($id = null) {
-		$vessel = $this->Vessels->get($id, [
-			'contain' => ['Creators', 'Modifiers', 'VesselOwners']
-		]);
-		$this->set('vessel', $vessel);
+		if($this->request->params['_ext']){
+			$vessel = $this->Vessels->get($id, [
+			//	'fields' => ['id','name','vessel_owner_id']
+			]);
+			$this->set('vessel', $vessel);
+	        $this->set('_serialize', ['vessel']);
+		}
+		else {
+			$vessel = $this->Vessels->get($id, [
+				'contain' => ['Creators', 'Modifiers', 'VesselOwners']
+			]);
+			$this->set('vessel', $vessel);
+		}
 	}
 
 /**
@@ -55,19 +66,35 @@ class VesselsController extends AppController {
  * @return void
  */
 	public function add() {
-		$vessel = $this->Vessels->newEntity($this->request->data);
-		if ($this->request->is('post')) {
-			if ($this->Vessels->save($vessel)) {
-				$this->Flash->success('The vessel has been saved.');
-				return $this->redirect(['action' => 'index']);
-			} else {
-				$this->Flash->error('The vessel could not be saved. Please, try again.');
-			}
+		if($this->request->params['_ext']){
+	        $vessel = $this->Vessels->newEntity($this->request->data);
+	        if ($this->Vessels->save($vessel, ['validate' => false])) {
+	            $message = 'Saved';
+	        } else {
+	            $message = 'Error';
+	        }
+	        $this->set([
+	            'data' => $this->request->data,
+	            'message' => $message,
+	            'vessel' => $vessel,
+	            '_serialize' => ['message', 'vessel', 'data']
+	        ]);
 		}
-		$creators = $this->Vessels->Creators->find('list');
-		$modifiers = $this->Vessels->Modifiers->find('list');
-		$vesselOwners = $this->Vessels->VesselOwners->find('list');
-		$this->set(compact('vessel', 'creators', 'modifiers', 'vesselOwners'));
+		else {
+			$vessel = $this->Vessels->newEntity($this->request->data);
+			if ($this->request->is('post')) {
+				if ($this->Vessels->save($vessel)) {
+					$this->Flash->success('The vessel has been saved.');
+					return $this->redirect(['action' => 'index']);
+				} else {
+					$this->Flash->error('The vessel could not be saved. Please, try again.');
+				}
+			}
+			$creators = $this->Vessels->Creators->find('list');
+			$modifiers = $this->Vessels->Modifiers->find('list');
+			$vesselOwners = $this->Vessels->VesselOwners->find('list');
+			$this->set(compact('vessel', 'creators', 'modifiers', 'vesselOwners'));
+		}
 	}
 
 /**
@@ -78,22 +105,43 @@ class VesselsController extends AppController {
  * @throws \Cake\Network\Exception\NotFoundException
  */
 	public function edit($id = null) {
-		$vessel = $this->Vessels->get($id, [
-			'contain' => []
-		]);
-		if ($this->request->is(['patch', 'post', 'put'])) {
-			$vessel = $this->Vessels->patchEntity($vessel, $this->request->data);
-			if ($this->Vessels->save($vessel)) {
-				$this->Flash->success('The vessel has been saved.');
-				return $this->redirect(['action' => 'index']);
-			} else {
-				$this->Flash->error('The vessel could not be saved. Please, try again.');
+		if($this->request->params['_ext']){
+			$vessel = $this->Vessels->get($id, [
+				'contain' => []
+			]);
+			if ($this->request->is(['patch', 'post', 'put'])) {
+				$vessel = $this->Vessels->patchEntity($vessel, $this->request->data);
+				if ($this->Vessels->save($vessel, ['validate' => false])) {
+	                $message = 'Saved';
+				} else {
+		            $message = 'Error';
+				}
 			}
+	        $this->set([
+	        	'vessel' => $vessel,
+	            'message' => $message,
+	            'data' => $this->request->data,
+	            '_serialize' => ['message','vessel','data']
+	        ]);
 		}
-		$creators = $this->Vessels->Creators->find('list');
-		$modifiers = $this->Vessels->Modifiers->find('list');
-		$vesselOwners = $this->Vessels->VesselOwners->find('list');
-		$this->set(compact('vessel', 'creators', 'modifiers', 'vesselOwners'));
+		else {
+			$vessel = $this->Vessels->get($id, [
+				'contain' => []
+			]);
+			if ($this->request->is(['patch', 'post', 'put'])) {
+				$vessel = $this->Vessels->patchEntity($vessel, $this->request->data);
+				if ($this->Vessels->save($vessel, ['validate' => false])) {
+					$this->Flash->success('The vessel has been saved.');
+					return $this->redirect(['action' => 'index']);
+				} else {
+					$this->Flash->error('The vessel could not be saved. Please, try again.');
+				}
+			}
+			$creators = $this->Vessels->Creators->find('list');
+			$modifiers = $this->Vessels->Modifiers->find('list');
+			$vesselOwners = $this->Vessels->VesselOwners->find('list');
+			$this->set(compact('vessel', 'creators', 'modifiers', 'vesselOwners'));
+		}
 	}
 
 /**
@@ -104,13 +152,26 @@ class VesselsController extends AppController {
  * @throws \Cake\Network\Exception\NotFoundException
  */
 	public function delete($id = null) {
-		$vessel = $this->Vessels->get($id);
-		$this->request->allowMethod(['post', 'delete']);
-		if ($this->Vessels->delete($vessel)) {
-			$this->Flash->success('The vessel has been deleted.');
-		} else {
-			$this->Flash->error('The vessel could not be deleted. Please, try again.');
+		if($this->request->params['_ext']){
+	        $vessel = $this->Vessels->get($id);
+	        $message = 'Deleted';
+	        if (!$this->Vessels->delete($vessel)) {
+	            $message = 'Error';
+	        }
+	        $this->set([
+	            'message' => $message,
+	            '_serialize' => ['message']
+	        ]);
 		}
-		return $this->redirect(['action' => 'index']);
+		else {		
+			$vessel = $this->Vessels->get($id);
+			$this->request->allowMethod(['post', 'delete']);
+			if ($this->Vessels->delete($vessel)) {
+				$this->Flash->success('The vessel has been deleted.');
+			} else {
+				$this->Flash->error('The vessel could not be deleted. Please, try again.');
+			}
+			return $this->redirect(['action' => 'index']);
+		}
 	}
 }
